Move RomfsEntry class
