+++ src/hactoolnet/ProcessRomfs.cs
@@ -1,5 +1,4 @@
 ﻿using System.IO;
-using LibHac;
 using LibHac.IO;
 
 namespace hactoolnet
